chore: Add NEXT_PUBLIC_INTERCOM_APP_ID to Dockerfile (#3330)
The Dockerfile has been updated to include the
`NEXT_PUBLIC_INTERCOM_APP_ID` environment variable. This change ensures
that the Intercom app ID is properly configured in the frontend
application.

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context.

## Checklist before requesting a review

Please delete options that are not relevant.

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my code
- [ ] I have commented hard-to-understand areas
- [ ] I have ideally added tests that prove my fix is effective or that
my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged

## Screenshots (if appropriate):