feat: Improve user credit calculation in get_user_credits (#3367)
Refactor the get_user_credits method in the Users class to improve the
calculation of monthly chat credits for a user. This change ensures that
the user's monthly chat credit is correctly deducted based on their
usage. If no user settings or monthly chat credit is found, a default
value of 25 is returned.

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context.

## Checklist before requesting a review

Please delete options that are not relevant.

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my code
- [ ] I have commented hard-to-understand areas
- [ ] I have ideally added tests that prove my fix is effective or that
my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged

## Screenshots (if appropriate):